move `close_button` to `ui/widgets.rs` as `x_button`
